Evaluating the Balance of Covariates After Matching

The function balance computes the standardized mean differences and the ratio of the variances among treatment groups, before and after matching. The function computes the two measures of balance for each pair of treatment groups.

Usage

balance(
  formulaBalance,
  match_id,
  data,
  weights_before = NULL,
  weights_after = NULL
)

Arguments

formulaBalance

Formula with form group ~ x_1 + ... + x_p. group is the variable identifying the treatment groups/exposures. The balance is evaluated for the covariates x_1,...,x_p. Numeric and integer variables are treated as continuous. Factor variables are treated as categorical. Factor variables with two levels are treated as binary.

match_id

Vector identifying the matched sets—matched units must have the same identifier. It is generated by polymatch.

data

The data.frame object with the data.

weights_before

Optional vector of weights of the observations to be considered in the unmatched dataset. To compute the unweighted standardized mean differences, set weights_before to NULL (default).

weights_after

Vector of weights for the matched dataset. Set it to NULL (default) to compute the unweighted standardized mean differences.

Value

A data.frame containing the standardized differences and ratios of the variances (only for continuous variables) for each pair of treatment groups. A graphical representation of the results can be generated with plotBalance.

See also

Examples

#Generate a datasets with group indicator and four variables:
#- var1, continuous, sampled from normal distributions;
#- var2, continuous, sampled from beta distributions;
#- var3, categorical with 4 levels;
#- var4, binary.
set.seed(1234567)
dat <- data.frame(group = c(rep("A",20),rep("B",60),rep("C",60)),
                  var1 = c(rnorm(20,mean=0,sd=1),
                           rnorm(60,mean=1,sd=2),
                           rnorm(60,mean=-1,sd=2)),
                  var2 = c(rbeta(20,shape1=1,shape2=1),
                           rbeta(60,shape1=2,shape2=1),
                           rbeta(60,shape1=1,shape2=2)),
                  var3 = factor(c(rbinom(20,size=3,prob=.4),
                                  rbinom(60,size=3,prob=.5),
                                  rbinom(60,size=3,prob=.3))),
                  var4 = factor(c(rbinom(20,size=1,prob=.5),
                                  rbinom(60,size=1,prob=.3),
                                  rbinom(60,size=1,prob=.7))))

#Match on propensity score
#-------------------------

#With multiple groups, need a multinomial model for the PS
library(VGAM)
#> Loading required package: stats4
#> Loading required package: splines
psModel <- vglm(group ~ var1 + var2 + var3 + var4,
                family=multinomial, data=dat)
#Estimated probabilities - 3 for each unit: P(group=A), P(group=B), P(group=C)
probsPS <- predict(psModel, type = "response")
dat$probA <- probsPS[,"A"]
dat$probB <- probsPS[,"B"]
dat$probC <- probsPS[,"C"]
#Estimated logits - 2 for each unit: log(P(group=A)/P(group=C)), log(P(group=B)/P(group=C))
logitPS <- predict(psModel, type = "link")
dat$logit_AvsC <- logitPS[,1]
dat$logit_BvsC <- logitPS[,2]

#Match on logits of PS
resultPs <- polymatch(group ~ logit_AvsC + logit_BvsC, data = dat,
                    distance = "euclidean")
#> Conditional optimal matching algorithm
#> Number of observations:  140 
#> Number of groups:  3 
#> Total distance of starting matched sample:  37.878 
#> Ended iteration  1  - total distance: 36.229 
#> End 
#> Number of iterations: 2, total distance:36.229
#> Number of matched sets:  20 
dat$match_id_ps <- resultPs$match_id

#Evaluate balance in covariates
tabBalancePs <- balance(group ~ var1 + var2 + var3 + var4,
                        match_id = dat$match_id_ps, data = dat)
tabBalancePs
#>    groups variable        type stdzDiffPre ratioVarsPre stdzDiffPost
#> 1     A-B     var1  continuous -0.51757085     4.425458   0.18521887
#> 2     A-C     var1  continuous  0.70879687     4.331368  -0.09488064
#> 3     B-C     var1  continuous  0.96044832     1.021723  -0.21982348
#> 4     A-B     var2  continuous -0.31012062     1.089382  -0.05870530
#> 5     A-C     var2  continuous  0.88429601     1.167713   0.15711852
#> 6     B-C     var2  continuous  1.22693868     1.071905   0.22174525
#> 7     A-B     var3 categorical  1.24193881           NA   0.82499540
#> 8     A-C     var3 categorical  0.41021435           NA   0.84149879
#> 9     B-C     var3 categorical  1.51855778           NA   0.24408401
#> 10    A-B     var4      binary  0.79026333           NA   0.43105272
#> 11    A-C     var4      binary -0.03514727           NA   0.21088361
#> 12    B-C     var4      binary -0.83129513           NA  -0.21685960
#>    ratioVarsPost
#> 1       2.173137
#> 2       4.184995
#> 3       1.925786
#> 4       1.121038
#> 5       1.003387
#> 6       1.117254
#> 7             NA
#> 8             NA
#> 9             NA
#> 10            NA
#> 11            NA
#> 12            NA

#You can also represent the standardized mean differences with 'plotBalance'
#plotBalance(tabBalancePs, ratioVariances = TRUE)